3: refactor away from emoji
