+++ qna-api/src/api.js
@@ -6,7 +6,7 @@ const cors = require('cors');
 require('dotenv').config();
 
 const middlewares = require('./middlewares');
-const api = require('./api');
+const userRouter = require('./routes/user.route');
 
 const app = express();
 
@@ -15,13 +15,7 @@ app.use(helmet());
 app.use(cors());
 app.use(express.json());
 
-app.get('/', (req, res) => {
-  res.json({
-    message: '🦄🌈✨👋🌎🌍🌏✨🌈🦄'
-  });
-});
-
-app.use('/api/v1', api);
+app.use('/api/v1/users', userRouter);
 
 app.use(middlewares.notFound);
 app.use(middlewares.errorHandler);
