+++ qna-api/src/controller/user.controller.js
@@ -1,0 +1,1 @@
+module.exports.getAllUsers = () => ['Tom', 'Dick', 'Harry'];
