+++ qna-api/src/index.js
@@ -1,4 +1,4 @@
-const app = require('./app');
+const app = require('./api');
 
 const port = process.env.PORT || 5000;
 app.listen(port, () => {
