+++ qna-api/src/routes/user.route.js
@@ -1,9 +1,11 @@
 const express = require('express');
 
+const userController = require('../controller/user.controller');
+
 const router = express.Router();
 
 router.get('/', (req, res) => {
-  res.json(['😀', '😳', '🙄']);
+  res.json(userController.getAllUsers());
 });
 
 module.exports = router;
